refactor: move bridge block type class to static folder
