+++ src/BridgeMaker.js
@@ -1,4 +1,4 @@
-const BRIDGE_BLOCK_TYPE = require('./BridgeBlockType');
+const BRIDGE_BLOCK_TYPE = require('./Static/BridgeBlockType');
 /**
  * 다리의 길이를 입력 받아서 다리를 생성해주는 역할을 한다.
  */
